+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## 1.15.4 - Unreleased
 ### Changed
 - Improved log messages of failed sendouts ([#138](https://github.com/putyourlightson/craft-campaign/issues/138)).
-
-### Changed
 - Improved the error message when trying to import a CSV file without selecting a mailing list ([#141](https://github.com/putyourlightson/craft-campaign/issues/141)).
+
+### Fixed
+- Fixed a bug that could throw the wrong exception when a pending contact is not found when verifying a subscription ([#149](https://github.com/putyourlightson/craft-campaign/issues/149)).
 
 ## 1.15.3 - 2020-05-12
 ### Fixed