+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release Notes for Campaign
 
-## 1.15.4 - Unreleased
+## 1.15.4 - 2020-06-18
+### Added
+- Added the `maxSendFailsAllowed` config setting which defines the failed attempts to send to contacts that are allowed before failing the entire sendout and defaults to 1 ([#138](https://github.com/putyourlightson/craft-campaign/issues/138)).
+
 ### Changed
 - Improved log messages of failed sendouts ([#138](https://github.com/putyourlightson/craft-campaign/issues/138)).
 - Improved the error message when trying to import a CSV file without selecting a mailing list ([#141](https://github.com/putyourlightson/craft-campaign/issues/141)).